extract Type Rendering
